if world rank is 0

+++ QC_pipeline/GPAW/mol_calc.py
@@ -2,7 +2,7 @@ from gpaw import restart,Davidson,RMMDIIS,Mixer,CG
 from ase.db import connect
 import QC_pipeline.GPAW.optimizer as opt
 import numpy as np
-from ase.parallel import paropen, parprint
+from ase.parallel import paropen, parprint, world
 from ase.io import read
 import os
 
@@ -54,7 +54,8 @@ class GPAW_mol_calculator:
                 raise RuntimeError('No HOMO LUMO Calculator.')
             self.atoms.set_calculator(calculator)
             opt.SPE_calc(self.atoms,name=cid+'/'+'homo-lumo'+'/'+file_name+'_occupied')
-            os.remove(file_prev+'.gpw')
+            if world.rank == 0:
+                os.remove(file_prev+'.gpw')
         elif mode == 'add_bands':
             file_prev='results/'+cid+'/'+'homo-lumo'+'/'+file_name+'_occupied'
             nbands=nbands_finder(file_prev+'.txt')
@@ -64,7 +65,8 @@ class GPAW_mol_calculator:
             #     convergence_criteria['bands']=nbands+add_convergence_bands
             self.atoms, calculator = restart(file_prev+'.gpw',nbands=int(nbands*bands_multiplier))
             self.file_dir_name=opt.SPE_calc(self.atoms,name=cid+'/'+'homo-lumo'+'/'+file_name+'_add_bands')
-            os.remove(file_prev+'.gpw')
+            if world.rank == 0:
+                os.remove(file_prev+'.gpw')
         elif mode == 'unoccupied':
             file_prev='results/'+cid+'/'+'homo-lumo'+'/'+file_name+'_add_bands'
             # eigen_arr=aboveLUMO_finder(file_prev+'.txt')
@@ -84,7 +86,8 @@ class GPAW_mol_calculator:
             self.atoms.set_calculator(calc_bands)
             self.file_dir_name=opt.SPE_calc(self.atoms,name=cid+'/'+'homo-lumo'+'/'+file_name+'_unoccupied',save_gpw=False)
             self.database_save('HOLO_'+file_name,option='homo-lumo')
-            os.remove(file_prev+'.gpw')
+            if world.rank == 0:
+                os.remove(file_prev+'.gpw')
         else:
             raise NameError('mode not definied. Available modes: occupied, add_bands, unoccupied.')
